id 자료형 수정 & Fetch type 수정.

+++ src/main/kotlin/sogong/sogongSpring/entity/HashtagDbEntity.kt
@@ -9,7 +9,7 @@ import javax.persistence.*
 class HashtagDbEntity(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private val hashid : Int,
+    private val hashid : Long,
 
     @Column(nullable = false, length = 45)
     private val hashname : String
