+++ src/main/kotlin/sogong/sogongSpring/repository/BestPostRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.BestPostEntity
 
-interface BestPostRepository : JpaRepository<BestPostEntity, Int> {
+interface BestPostRepository : JpaRepository<BestPostEntity, Long> {
 }
