+++ src/main/kotlin/sogong/sogongSpring/repository/EntireCommentRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.EntireCommentEntity
 
-interface EntireCommentRepository : JpaRepository<EntireCommentEntity, Int> {
+interface EntireCommentRepository : JpaRepository<EntireCommentEntity, Long> {
 }
